+++ sdk/eventhubs/azure-messaging-eventhubs/src/main/java/com/azure/messaging/eventhubs/SystemProperties.java
@@ -25,11 +25,12 @@ import java.util.stream.Collectors;
  */
 final class SystemProperties implements Map<String, Object> {
     private static final ClientLogger LOGGER = new ClientLogger(SystemProperties.class);
-    private final Long offset;
+    private final String offset;
     private final String partitionKey;
     private final Instant enqueuedTime;
     private final Long sequenceNumber;
     private final AmqpAnnotatedMessage message;
+    private final Integer replicationSegment;
 
     /**
      * Creates an empty set of system properties. This is the case where a message was not received.
@@ -40,15 +41,17 @@ final class SystemProperties implements Map<String, Object> {
         this.enqueuedTime = null;
         this.partitionKey = null;
         this.sequenceNumber = null;
+        this.replicationSegment = null;
     }
 
-    SystemProperties(final AmqpAnnotatedMessage message, long offset, Instant enqueuedTime, long sequenceNumber,
-        String partitionKey) {
+    SystemProperties(final AmqpAnnotatedMessage message, String offset, Instant enqueuedTime, long sequenceNumber,
+        String partitionKey, Integer replicationSegment) {
         this.message = Objects.requireNonNull(message, "'message' cannot be null.");
         this.offset = offset;
         this.enqueuedTime = enqueuedTime;
         this.sequenceNumber = sequenceNumber;
         this.partitionKey = partitionKey;
+        this.replicationSegment = replicationSegment;
     }
 
     /**
@@ -56,7 +59,7 @@ final class SystemProperties implements Map<String, Object> {
      *
      * @return The offset within the Event Hubs stream.
      */
-    Long getOffset() {
+    String getOffset() {
         return offset;
     }
 
@@ -90,6 +93,15 @@ final class SystemProperties implements Map<String, Object> {
      */
     Long getSequenceNumber() {
         return sequenceNumber;
+    }
+
+    /**
+     * Gets the replication segment for this event. -1 or null if geo-disaster recovery is not enabled.
+     *
+     * @return The replication segment for this event. -1 or null of geo-disaster recovery is not enabled.
+     */
+    Integer getReplicationSegment() {
+        return replicationSegment;
     }
 
     @Override