+++ sdk/eventhubs/azure-messaging-eventhubs-checkpointstore-blob/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Release History
 
+## 1.20.0-beta.1 (Unreleased)
+
+### Features Added
+
+- Adds support for persisting replication segment in `Checkpoint`.
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+
 ## 1.19.3 (2024-04-23)
 
 ### Other Changes