+++ sdk/eventhubs/azure-messaging-eventhubs/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Release History
 
+## 5.19.0-beta.1 (Unreleased)
+
+### Features Added
+
+- Added support for Geo Disaster recovery.
+- Added `EventData.getReplicationSegment()`.
+- Added `getBeginningEnqueuedReplicationSegment()` and `getLastEnqueuedReplicationSegment()` to `PartitionProperties`.
+- Added `getReplicationSegment()` into `LastEnqueuedEventProperties`.
+
+### Breaking Changes
+
+- `EventData.getOffset()`, `CheckpointStore.getOffset()`, `EventData.getOffset(), and `LastEnqueuedEventProperties.getOffset()` are changed from `Long` to `String`.
+
+### Bugs Fixed
+
+- Fixed issue where creating EventProcessorClient instances using the same EventProcessorClientBuilder instance could result in incorrect properties. ([#29875](https://github.com/Azure/azure-sdk-for-java/issues/29875))
+
+### Other Changes
+
+
 ## 5.18.3 (2024-04-23)
 
 ### Other Changes